Move main plugin to index
