+++ src/exports/plugins/index.js
@@ -1,3 +1,2 @@
 export { dev_server_config } from './dev-server-config.js';
 export { dev_server_manifest } from './dev-server-manifest.js';
-export { v4wp } from './v4wp.js';
